+++ lib/OrdinaryDiffEqSymplecticRK/Project.toml
@@ -1,0 +1,4 @@
+name = "OrdinaryDiffEqSymplecticRK"
+uuid = "fa646aed-7ef9-47eb-84c4-9443fc8cbfa8"
+authors = ["ParamThakkar123 <paramthakkar864@gmail.com>"]
+version = "0.1.0"
