+++ lib/OrdinaryDiffEqSymplecticRK/src/alg_utils.jl
@@ -1,0 +1,17 @@
+alg_order(alg::SymplecticEuler) = 1
+alg_order(alg::VelocityVerlet) = 2
+alg_order(alg::VerletLeapfrog) = 2
+alg_order(alg::PseudoVerletLeapfrog) = 2
+alg_order(alg::McAte2) = 2
+alg_order(alg::Ruth3) = 3
+alg_order(alg::McAte3) = 3
+alg_order(alg::McAte4) = 4
+alg_order(alg::CandyRoz4) = 4
+alg_order(alg::CalvoSanz4) = 4
+alg_order(alg::McAte42) = 4
+alg_order(alg::McAte5) = 5
+alg_order(alg::Yoshida6) = 6
+alg_order(alg::KahanLi6) = 6
+alg_order(alg::McAte8) = 8
+alg_order(alg::KahanLi8) = 8
+alg_order(alg::SofSpa10) = 10
