+++ lib/OrdinaryDiffEqSymplecticRK/test/runtests.jl
@@ -1,0 +1,3 @@
+using SafeTestsets
+
+@time @safetestset "Synplectic Tests" include("symplectic_tests.jl")
